requirements.txt
Flask==3.0.3
python-whois==0.9.4
requests==2.32.3
